Renamed and refactored DisplayData to SchemaDisplay and DisplaySchemaContainer
